pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "context-scribe"
version = "0.1.0"
description = "A persistent secretary daemon to eliminate Agent Amnesia"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "watchdog",
    "rich",
    "mcp",
    "click",
    "PyYAML"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "pytest-cov"
]
dev = [
    "ruff"
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W"]  # PEP8 equivalent: errors, pyflakes, warnings

[project.scripts]
context-scribe = "context_scribe.main:cli"
